# Faster From<[u8]> for BytesMut, remove panic in fmt::Write