Benchmark change for `From<[u8]>`:

Before:
```
test from_long_slice      ... bench:       45 ns/iter (+/- 2) = 2844 MB/s
```
After:
```
test from_long_slice      ... bench:       40 ns/iter (+/- 2) = 3200 MB/s
```