@seanmonstar 2017-06-15
The behavior is now the same as `&mut [u8]`, where if the string won't fit, a `fmt::Error` is returned.